Rename parser to transformer
