+++ src/printer/fetchFns/body/printFnBody.ts
@@ -1,4 +1,4 @@
-import type { ExtractedOperation } from "../../../parser/buildOperations.js";
+import type { ExtractedOperation } from "../../../transformer/operations/buildOperations.js";
 import { printOptions } from "./printOptions.js";
 import { printReturn } from "./printReturn.js";
 import { printUrl } from "./printUrl.js";
