+++ src/printer/fetchFns/body/printUrl.ts
@@ -1,4 +1,4 @@
-import type { ExtractedOperation } from "../../../parser/buildOperations.js";
+import type { ExtractedOperation } from "../../../transformer/operations/buildOperations.js";
 
 export const printUrl = ({ path, parameterTypes }: ExtractedOperation) => {
   const replacedPath = replacePathParams(path);
